requirements.txt
Flask
gunicorn
SQLAlchemy
python-dotenv
pymysql
